added paste operation
