+++ libs/designer/src/lib/ui/panel/recommendation/recommendationPanelContext.tsx
@@ -1,4 +1,4 @@
-import { useNodeDisplayName, useNodeMetadata, type AppDispatch } from '../../../core';
+import { useNodeDisplayName, type AppDispatch } from '../../../core';
 import { addOperation } from '../../../core/actions/bjsworkflow/add';
 import { useAllConnectors, useAllOperations } from '../../../core/queries/browse';
 import { useHostOptions } from '../../../core/state/designerOptions/designerOptionsSelectors';
@@ -17,7 +17,7 @@ import { SearchView } from './searchView';
 import { Link, Icon } from '@fluentui/react';
 import { Button } from '@fluentui/react-components';
 import { bundleIcon, Dismiss24Filled, Dismiss24Regular } from '@fluentui/react-icons';
-import { SearchService, equals, guid, areApiIdsEqual, removeIdTag } from '@microsoft/logic-apps-shared';
+import { SearchService, equals, guid, areApiIdsEqual } from '@microsoft/logic-apps-shared';
 import { Card, OperationSearchHeader, XLargeText } from '@microsoft/designer-ui';
 import type { CommonPanelProps } from '@microsoft/designer-ui';
 import type { DiscoveryOpArray, DiscoveryOperation, DiscoveryResultTypes } from '@microsoft/logic-apps-shared';
@@ -26,9 +26,8 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useIntl } from 'react-intl';
 import { useDispatch } from 'react-redux';
 import { retrieveClipboardData } from '../../../core/utils/clipboard';
-import { pasteOperation, pasteScopeOperation, type PasteOperationPayload } from '../../../core/actions/bjsworkflow/copypaste';
-import { useEdgeContextMenuData } from '../../../core/state/designerView/designerViewSelectors';
-import { useUpstreamNodes } from '../../../core/state/tokens/tokenSelectors';
+import type { PasteOperationPayload } from '../../../core/actions/bjsworkflow/copypaste';
+import { PasteOperation } from '../../../ui/common/OperationsCopyPaste/pasteOperation';
 
 const CloseIcon = bundleIcon(Dismiss24Filled, Dismiss24Regular);
 
@@ -226,61 +225,6 @@ export const RecommendationPanelContext = (props: CommonPanelProps) => {
   }, []);
   const displayName = useNodeDisplayName(copiedNode?.nodeId);
 
-  const menuData = useEdgeContextMenuData();
-  const graphId = useMemo(() => menuData?.graphId, [menuData]);
-  const parentId = useMemo(() => menuData?.parentId, [menuData]);
-  const childId = useMemo(() => menuData?.childId, [menuData]);
-  const nodeMetadata = useNodeMetadata(removeIdTag(parentId ?? ''));
-
-  const newParentId = useMemo(() => {
-    if (nodeMetadata?.subgraphType) {
-      return nodeMetadata.parentNodeId;
-    }
-    return parentId;
-  }, [nodeMetadata, parentId]);
-
-  const upstreamNodesOfChild = useUpstreamNodes(removeIdTag(childId ?? newParentId ?? graphId ?? ''));
-
-  const handlePasteClicked = useCallback(async () => {
-    if (!graphId) {
-      return;
-    }
-    const relationshipIds = { graphId, childId, parentId };
-    const copiedNode = await retrieveClipboardData();
-    if (!copiedNode) {
-      return;
-    }
-    if (copiedNode?.isScopeNode) {
-      dispatch(
-        pasteScopeOperation({
-          relationshipIds,
-          nodeId: copiedNode.nodeId,
-          serializedValue: copiedNode.serializedOperation,
-          allConnectionData: copiedNode.allConnectionData,
-          staticResults: copiedNode.staticResults,
-          upstreamNodeIds: upstreamNodesOfChild,
-        })
-      );
-    } else {
-      dispatch(
-        pasteOperation({
-          relationshipIds,
-          nodeId: copiedNode.nodeId,
-          nodeData: copiedNode.nodeData,
-          nodeTokenData: copiedNode.nodeTokenData,
-          operationInfo: copiedNode.nodeOperationInfo,
-          connectionData: copiedNode.nodeConnectionData,
-          comment: copiedNode.nodeComment,
-        })
-      );
-    }
-    // LoggerService().log({
-    //   area: 'EdgeContextualMenu:handlePasteClicked',
-    //   level: LogEntryLevel.Verbose,
-    //   message: 'New node added via paste.',
-    // });
-  }, [graphId, childId, parentId, dispatch, upstreamNodesOfChild]);
-
   return (
     <>
       {copiedNode && (
@@ -289,16 +233,17 @@ export const RecommendationPanelContext = (props: CommonPanelProps) => {
             <XLargeText text={pasteActionHeadingText} />
             <Button aria-label={closeButtonAriaLabel} appearance="subtle" onClick={toggleCollapse} icon={<CloseIcon />} />
           </div>
-          <Card
-            title={displayName}
-            brandColor={copiedNode?.nodeData?.operationMetadata?.brandColor}
-            drag={undefined}
-            draggable={false}
-            dragPreview={undefined}
-            icon={copiedNode?.nodeData?.operationMetadata?.iconUri}
-            id={copiedNode.nodeId}
-            onClick={handlePasteClicked}
-          />
+          <PasteOperation location="RecommendationPanel">
+            <Card
+              title={displayName}
+              brandColor={copiedNode?.nodeData?.operationMetadata?.brandColor}
+              drag={undefined}
+              draggable={false}
+              dragPreview={undefined}
+              icon={copiedNode?.nodeData?.operationMetadata?.iconUri}
+              id={copiedNode.nodeId}
+            />
+          </PasteOperation>
         </>
       )}
       <div className="msla-app-action-header">
